refactor(ddd): contrôleurs IA sous app/AI/Interfaces + commandes EdgeSync dans le module (Closes #6578)
- 5 contrôleurs IA (Agent, AIAnalytics, AIGateway, ConversationExport, Voice)
  déplacés app/Http/Controllers/AI/ → app/AI/Interfaces/Api/V1/Controllers/
  (namespace App\AI\Interfaces\Api\V1\Controllers) — routes ai.php/api.php
  + test TtsExecSecurityTest repointés.
- 2 commandes EdgeSync (EdgeSyncDaemonCommand, MonitorEdgeNodesCommand)
  déplacées app/Console/Commands/ → Modules/EdgeSync/Console/Commands/ et
  enregistrées via ->commands() dans EdgeSyncServiceProvider (pattern
  Accounting).
- ARCHITECTURE.md : dérogation documentée pour app/Http/Controllers/Web/
  (surface session/blade distincte de l'API JSON par module).
