+++ api/app/AI/Interfaces/Api/V1/Controllers/ConversationExportController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers\AI;
+namespace App\AI\Interfaces\Api\V1\Controllers;
 
 use App\AI\Models\AiExport;
 use App\AI\Services\ConversationExportService;
